Build an addition calculator and use the method to_i and to_s

+++ basics.rb
@@ -1,5 +1,5 @@
 puts "Enter number 1"
-num1 = gets.chomp().to_i
+num1 = gets.chomp().to_f
 puts "Enter number 2"
 num2 = gets.chomp().to_i
